feat: small code clean up and rename HttpMethod values
