+++ networking/src/main/java/com/kmc/networking/HttpMethod.kt
@@ -1,0 +1,5 @@
+package com.kmc.networking
+
+enum class HttpMethod {
+    GET, POST, PUT, PATCH, DELETE
+}
