--- networking/src/main/java/com/kmc/networking/entity/HttpMethod.kt
@@ -1,5 +1,0 @@
-package com.kmc.networking.entity
-
-enum class HttpMethod {
-    Get, Post, Put, Patch, Delete
-}
